requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
openai
python-dotenv
streamlit
